<chore> connection error
- config.py 와 .env 파일의 위치를 app 안으로 변경하였습니다.

close #80
